Finished moving styles to respective folders
