+++ resources/containers/auth-button.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { useTranslations } from "next-intl";
-import form_style from "@/styles/forms/form.module.css";
+import form_style from "./styles/auth-button.module.css";
 
 interface AuthButtonProps {
   action: "register" | "login";
@@ -10,7 +10,7 @@ const AuthButton: React.FC<AuthButtonProps> = ({ action }) => {
   const t = useTranslations("");
 
   return (
-    <div className={form_style.buttonWrapper}>
+    <div>
       <button type="submit" className={form_style.submitButton}>
         {action === "register" ? t("Register") : t("Login")}
       </button>
