+++ resources/containers/styles/auth-button.module.css
@@ -1,0 +1,15 @@
+.submitButton {
+  background: #1C364B;
+  color: white;
+  padding: 0.5rem 1rem;
+  border: none;
+  border-radius: 4px;
+  cursor: pointer;
+  width: 100%;
+  max-width: 200px;
+  margin-bottom: 1rem;
+  margin-top: 1rem;
+
+  font-family: 'Montserrat', sans-serif;
+  font-size: 0.8rem;
+}
